Add ServiceClientAdmin file and dependencies
