+++ revashare-svc-webapi/revashare-svc-webapi.Logic/Interfaces/IAdmin.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace revashare_svc_webapi.Logic.Interfaces
+{
+    public interface IAdmin
+    {
+
+    }
+}
